+++ cmake/modules/DunePythonCommonMacros.cmake
@@ -72,6 +72,8 @@ include(DunePythonTestCommand)
 # Find the Python Interpreter and libraries
 find_package(Python3 COMPONENTS Interpreter Development)
 
+set(HAVE_PYTHON3_EMBED ${Python3_Development.Embed_FOUND})
+
 if(Python3_Interpreter_FOUND)
   include(DuneExecuteProcess)
   include(DunePathHelper)