+++ dune/common/parametertreeparser.cc
@@ -17,6 +17,11 @@
 #include <map>
 #include <algorithm>
 
+#if HAVE_PYTHON3_EMBED
+#include <dune/python/pybind11/embed.h>
+namespace py = pybind11;
+#endif
+
 #include <dune/common/exceptions.hh>
 
 std::string Dune::ParameterTreeParser::ltrim(const std::string& s)
@@ -258,4 +263,52 @@ std::string Dune::ParameterTreeParser::generateHelpString(
         keywords[i] + ":\t" + help[i] + "\n";
   }
   return helpstr;
+}
+
+#if HAVE_PYTHON3_EMBED
+namespace {
+  void addPythonDict(const py::dict& scope,
+                     Dune::ParameterTree& pt,
+                     bool overwrite,
+                     std::vector<PyObject*>& ancestors){
+    for (const auto& [key, value] : scope){
+      const std::string& key_str = key.cast<std::string>();
+      // ignore entry if the key starts with __
+      if(key_str[0] == '_' && key_str[1] == '_')
+        continue;
+      // if the value is a dict create a sub-ParameterTree
+      if(py::isinstance<py::dict>(value)){
+        if (std::find(ancestors.begin(), ancestors.end(), value.ptr()) != ancestors.end()){
+          DUNE_THROW(Dune::Exception, "Cannot parse python dictionary as ParameterTree: Cycle detected!");
+        }
+        ancestors.push_back(value.ptr());
+        const py::dict& dict = value.cast<py::dict>();
+        addPythonDict(dict, pt.sub(key_str), overwrite, ancestors);
+        ancestors.pop_back();
+        continue;
+      }
+      // skip if overwrite is false and key is already contained in the ParameterTree
+      if(!overwrite and pt.hasKey(key_str))
+        continue;
+      // otherwise add the python string representation as value
+      std::string val_str = py::str(value);
+      pt[key_str] = val_str;
+    }
+  }
+}
+#endif
+
+void Dune::ParameterTreeParser::readPythonTree(std::string file,
+                                               ParameterTree& pt,
+                                               bool overwrite)
+{
+#if HAVE_PYTHON3_EMBED
+  py::scoped_interpreter guard{};
+  py::dict scope = py::module_::import("__main__").attr("__dict__");
+  py::eval_file(file);
+  std::vector<PyObject*> anchestors = {scope.ptr()};
+  addPythonDict(scope, pt, overwrite, anchestors);
+#else
+  DUNE_THROW(Dune::NotImplemented, "Python needs to be installed to parse a python script as ParameterTree.");
+#endif
 }